sampledata.mongo.js fixes (not fully tested)

+++ src/samples/sampleData.mongo.js
@@ -9,7 +9,7 @@ module.exports = (
     // pass true to reset from sampleData,
     // pass an {object} of key:data's to reset to that database
     // pass a 'key' to reset only that key from sampleData  
-    reset = true,
+    reset = false,
 
     // set to true to delete any dataset not represented in reset
     deleteWhenNotInReset = false
@@ -20,30 +20,33 @@ module.exports = (
     .then(async client => {
 
         let db = client.db();
-        let collections = await db.collections();
+        let collectionNames = await db.collections().map(c => c.s.name);
 
         if (!reset)
             return db;
 
-        let dataset = 
+        let datasets = 
             reset === true ? sampleDataSets
             : typeof reset === 'object' && Object.keys(reset).length > 0 ? reset
-            : typeof reset === 'string' ? sampleDataSets[reset]
+            : typeof reset === 'string' ? { [reset]: sampleDataSets[reset] }
             : null;
 
         let deleteKeys = 
-            deleteWhenNotInReset ? collections.map(c => c.s.name)
-            : Object.keys(dataset);
+            deleteWhenNotInReset 
+            ? collectionNames
+            : Object.keys(datasets);
 
         for (let key of deleteKeys) {                
+            if (collectionNames.indexOf(key) == -1)
+                continue;
             console.log('deleteing: ' + key);
             await db.dropCollection(key);
         }
 
-        for (let key of Object.keys(dataset)) {
+        for (let key of Object.keys(datasets)) {
             console.log('creating: ' + key);
             await db.createCollection(key); 
-            await db.collection(key).insertMany(dataset[key]);
+            await db.collection(key).insertMany(datasets[key]);
         }
 
         return db;
